Refactor models and update API constraints

+++ LelkiBekeAdmin/Models/ModifyMenuItem.cs
@@ -13,7 +13,7 @@ namespace LelkiBekeAdmin.Models
         public int category_id { get; set; }
         public string name { get; set; }
         public string description { get; set; }
-        public float price { get; set; }
+        public int price { get; set; }
         public string image_url { get; set; }
     }
