+++ LelkiBekeAdmin/Models/NewMenuItem.cs
@@ -11,7 +11,7 @@ namespace LelkiBekeAdmin.Models
         public int category_id { get; set; }
         public string name { get; set; }
         public string description { get; set; }
-        public float price { get; set; }
+        public int price { get; set; }
         public string image_url { get; set; }
     }
